Port custom color picker to Ace3 widget, necessary for 10.2.5 API change
While we really need to rewrite the whole GUI, for the moment we can just port our custom color widget to use Ace3's color widget. This is required due to the API changes that happened in 10.2.5. Ace3 will make this a lot easier to deal with the two different versions we need now between Classic and Retail