General code cleanups, and added some error messages to fix #11
If you tried to call :AddData() too early from another addon, there was no guarantee that DevTool was initialized enough to be ready for the incoming data. I moved a couple items earlier in the loading process to try to help this, and added some safety checks and error messaging to handle this cleaner than just erroring out.